+++ pkg/innerring/processors/netmap/process_peers.go
@@ -62,17 +62,6 @@ func (np *Processor) processUpdatePeer(ev netmapEvent.UpdatePeer) {
 
 	var err error
 
-	if ev.Maintenance() {
-		err = np.nodeStateSettings.MaintenanceModeAllowed()
-		if err != nil {
-			np.log.Info("prevent switching node to maintenance state",
-				zap.Error(err),
-			)
-
-			return
-		}
-	}
-
 	nr := ev.NotaryRequest()
 	err = np.netmapClient.Morph().NotarySignAndInvokeTX(nr.MainTransaction, false)